refactor(releases): clean up main directory and releases
