+++ src/Main/App.js
@@ -1,5 +1,5 @@
 import React, {Component, Fragment} from 'react';
-import HomeReleases from './HomeReleases';
+import HomeReleases from './Home/HomeReleases';
 import './App.css';
 
 class App extends Component {
