+++ src/index.js
@@ -1,11 +1,11 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
-import App from './App';
-import Header from './Header';
+import App from './Main/App';
+import Header from './Main/Header/Header';
 import history from './history'
 import Release0001 from './Releases/Release0001';
-import MorphingBalls from "./Releases/MorphingBalls";
+import MorphingBalls from "./SpikeWork/MorphingBalls";
 import Release0002 from "./Releases/Release0002";
 import Release0003 from "./Releases/Release0003";
 import Release0004 from "./Releases/Release0004/index";
