--- src/nav.js
@@ -1,5 +1,0 @@
-import history from './history'
-
-export default function nav(loc) {
-  history.push(loc);
-}
